@@ -2,8 +2,8 @@ from setuptools import find_packages, setup
 
 setup(
 	name='NEMO',
-	version='3.13.3',
-	python_requires='>=3.6',
+	version='3.14.0',
+	python_requires='>=3.7',
 	packages=find_packages(exclude=['NEMO.tests','NEMO.tests.*']),
 	include_package_data=True,
 	url='https://github.com/usnistgov/NEMO',
@@ -24,17 +24,17 @@ setup(
 		'Programming Language :: Python :: 3.6',
 	],
 	install_requires=[
-		'cryptography==3.4.7',
-		'Django==2.2.24',
-		'django-filter==2.4.0',
-		'drf-renderer-xlsx==0.4.1',
-		'djangorestframework==3.12.4',
-		'drf-flex-fields==0.9.3',
-		'ldap3==2.9',
-		'python-dateutil==2.8.1',
-		'requests==2.25.1',
-		'Pillow==8.3.2',
-		'django-mptt==0.12.0',
+		'cryptography==36.0.1',
+		'Django==2.2.26',
+		'django-filter==21.1',
+		'drf-renderer-xlsx==0.4.4',
+		'djangorestframework==3.13.1',
+		'drf-flex-fields==0.9.7',
+		'ldap3==2.9.1',
+		'python-dateutil==2.8.2',
+		'requests==2.27.1',
+		'Pillow==9.0.0',
+		'django-mptt==0.13.4',
 	],
 	entry_points={
 		'console_scripts': ['nemo=NEMO.provisioning:entry_point'],
