@@ -70,9 +70,9 @@ def staff_member_required(view_func=None, redirect_field_name=REDIRECT_FIELD_NAM
 	return actual_decorator
 
 
-def superuser_required(view_func=None, redirect_field_name=REDIRECT_FIELD_NAME, login_url=None):
+def administrator_required(view_func=None, redirect_field_name=REDIRECT_FIELD_NAME, login_url=None):
 	"""
-	Decorator for views that checks that the user is logged in and is a superuser.
+	Decorator for views that checks that the user is logged in and is an administrator.
 	"""
 	actual_decorator = user_passes_test(
 		lambda u: u.is_active and u.is_superuser, login_url=login_url, redirect_field_name=redirect_field_name
