@@ -14,18 +14,20 @@ from rest_framework import routers
 from NEMO.models import ReservationItemType
 from NEMO.views import (
 	abuse,
+	access_requests,
 	accounts_and_projects,
 	alerts,
 	api,
 	area_access,
 	authentication,
-	buddy_system,
+	buddy_requests,
 	calendar,
 	configuration_agenda,
 	consumables,
 	contact_staff,
 	customization,
 	email,
+	event_details,
 	feedback,
 	get_projects,
 	history,
@@ -46,6 +48,7 @@ from NEMO.views import (
 	training,
 	tutorials,
 	usage,
+	user_requests,
 	users,
 )
 
@@ -127,12 +130,22 @@ urlpatterns += [
 	url(r'^tool_usage_group_question/(?P<tool_id>\d+)/(?P<group_name>\w+)/$', tool_control.tool_usage_group_question, name='tool_usage_group_question'),
 	url(r'^reset_tool_counter/(?P<counter_id>\d+)/$', tool_control.reset_tool_counter, name='reset_tool_counter'),
 
+	# User requests
+	url(r'^user_requests/$', user_requests.user_requests, name='user_requests'),
+	url(r'^user_requests/(?P<tab>buddy|access)/$', user_requests.user_requests, name='user_requests'),
+
+	# Access requests
+	url(r'^access_requests/$', access_requests.access_requests, name='access_requests'),
+	url(r'^create_access_request/$', access_requests.create_access_request, name='create_access_request'),
+	url(r'^edit_access_request/(?P<request_id>\d+)/$', access_requests.create_access_request, name='edit_access_request'),
+	url(r'^delete_access_request/(?P<request_id>\d+)/$', access_requests.delete_access_request, name='delete_access_request'),
+
 	# Buddy System
-	url(r'^buddy_system/$', buddy_system.buddy_system, name='buddy_system'),
-	url(r'^create_buddy_request/$', buddy_system.create_buddy_request, name='create_buddy_request'),
-	url(r'^edit_buddy_request/(?P<request_id>\d+)/$', buddy_system.create_buddy_request, name='edit_buddy_request'),
-	url(r'^delete_buddy_request/(?P<request_id>\d+)/$', buddy_system.delete_buddy_request, name='delete_buddy_request'),
-	url(r'^buddy_request_reply/(?P<request_id>\d+)/$', buddy_system.buddy_request_reply, name='buddy_request_reply'),
+	url(r'^buddy_requests/$', buddy_requests.buddy_requests, name='buddy_requests'),
+	url(r'^create_buddy_request/$', buddy_requests.create_buddy_request, name='create_buddy_request'),
+	url(r'^edit_buddy_request/(?P<request_id>\d+)/$', buddy_requests.create_buddy_request, name='edit_buddy_request'),
+	url(r'^delete_buddy_request/(?P<request_id>\d+)/$', buddy_requests.delete_buddy_request, name='delete_buddy_request'),
+	url(r'^buddy_request_reply/(?P<request_id>\d+)/$', buddy_requests.buddy_request_reply, name='buddy_request_reply'),
 
 	# Tasks:
 	url(r'^create_task/$', tasks.create, name='create_task'),
@@ -155,12 +168,14 @@ urlpatterns += [
 	url(r'^cancel_outage/(?P<outage_id>\d+)/$', calendar.cancel_outage, name='cancel_outage'),
 	url(r'^set_reservation_title/(?P<reservation_id>\d+)/$', calendar.set_reservation_title, name='set_reservation_title'),
 	url(r'^change_reservation_project/(?P<reservation_id>\d+)/$', calendar.change_reservation_project, name='change_reservation_project'),
-	url(r'^event_details/reservation/(?P<reservation_id>\d+)/$', calendar.reservation_details, name='reservation_details'),
-	url(r'^event_details/outage/(?P<outage_id>\d+)/$', calendar.outage_details, name='outage_details'),
-	url(r'^event_details/usage/(?P<event_id>\d+)/$', calendar.usage_details, name='usage_details'),
-	url(r'^event_details/area_access/(?P<event_id>\d+)/$', calendar.area_access_details, name='area_access_details'),
 	url(r'^proxy_reservation/$', calendar.proxy_reservation, name='proxy_reservation'),
 	url(r'^reservation_group_question/(?P<reservation_question_id>\d+)/(?P<group_name>\w+)/$', calendar.reservation_group_question, name='reservation_group_question'),
+
+	# Event Details:
+	url(r'^event_details/reservation/(?P<reservation_id>\d+)/$', event_details.reservation_details, name='reservation_details'),
+	url(r'^event_details/outage/(?P<outage_id>\d+)/$', event_details.outage_details, name='outage_details'),
+	url(r'^event_details/usage/(?P<event_id>\d+)/$', event_details.usage_details, name='usage_details'),
+	url(r'^event_details/area_access/(?P<event_id>\d+)/$', event_details.area_access_details, name='area_access_details'),
 
 	# Qualifications:
 	url(r'^qualifications/$', qualifications.qualifications, name='qualifications'),
@@ -173,6 +188,7 @@ urlpatterns += [
 	url(r'^end_staff_charge/$', staff_charges.end_staff_charge, name='end_staff_charge'),
 	url(r'^begin_staff_area_charge/$', staff_charges.begin_staff_area_charge, name='begin_staff_area_charge'),
 	url(r'^end_staff_area_charge/$', staff_charges.end_staff_area_charge, name='end_staff_area_charge'),
+	url(r'^edit_staff_charge_note/$', staff_charges.edit_staff_charge_note, name='edit_staff_charge_note'),
 
 	# Status dashboard:
 	url(r'^status_dashboard/$', status_dashboard.status_dashboard, name='status_dashboard'),
@@ -275,7 +291,7 @@ urlpatterns += [
 	url(r'^media/(?P<path>.*)$', serve, {'document_root': settings.MEDIA_ROOT}, name='media'),
 
 	# User Preferences
-	url(r'^user_preferences/$', users.user_preferences, name='user_preferences')
+	url(r'^user_preferences/$', users.user_preferences, name='user_preferences'),
 ]
 
 if settings.ALLOW_CONDITIONAL_URLS:
@@ -295,6 +311,7 @@ if settings.ALLOW_CONDITIONAL_URLS:
 		url(r'^email_usage_reminders/$', calendar.email_usage_reminders, name='email_usage_reminders'),
 		url(r'^email_out_of_time_reservation_notification/$', calendar.email_out_of_time_reservation_notification, name='email_out_of_time_reservation_notification'),
 		url(r'^cancel_unused_reservations/$', calendar.cancel_unused_reservations, name='cancel_unused_reservations'),
+		url(r'^email_weekend_access_notification/$', access_requests.email_weekend_access_notification, name='email_weekend_access_notification'),
 
 		# Abuse:
 		url(r'^abuse/$', abuse.abuse, name='abuse'),
