@@ -1,4 +1,3 @@
-from datetime import datetime
 from logging import getLogger
 from typing import List, Set
 
@@ -22,7 +21,14 @@ from NEMO.models import (
 	UsageEvent,
 	User,
 )
-from NEMO.utilities import BasicDisplayTable, get_month_timeframe, month_list, parse_start_and_end_date
+from NEMO.utilities import (
+	BasicDisplayTable,
+	export_format_datetime,
+	format_datetime,
+	get_month_timeframe,
+	month_list,
+	parse_start_and_end_date,
+)
 from NEMO.views.api_billing import (
 	BillableItem,
 	billable_items_area_access_records,
@@ -352,9 +358,12 @@ def csv_export_response(usage_events, area_access, training_sessions, staff_char
 	data.extend(billable_items_area_access_records(area_access))
 	data.extend(billable_items_usage_events(usage_events))
 	for billable_item in data:
+		# Format start & end times
+		billable_item.start = format_datetime(billable_item.start, "SHORT_DATETIME_FORMAT")
+		billable_item.end = format_datetime(billable_item.end, "SHORT_DATETIME_FORMAT")
 		table_result.add_row(vars(billable_item))
 	response = table_result.to_csv()
-	filename = f"usage_export_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.csv"
+	filename = f"usage_export_{export_format_datetime()}.csv"
 	response["Content-Disposition"] = f'attachment; filename="{filename}"'
 	return response
 
